Add the built-in `Lazy` class
This class is made to work like the Python partial() function,
i.e. it receives a functions and a set of parameters and return
an object that is capable of execute that function with those
parameters.

In python the object returned a new function, the difference in
this case is that the object returned is an instance of
the Lazy class, and thus it allows the user to access the function
and the parameters if he wants. To execute the function the user
can call the `exec` function of the class.