Simplify representation of the  Matcher class
Before this commit hooks were an unique type, and were not meant
to interact with the rest of the language.

Now hooks are functions, and can be directly executed from
within the language.

Furthermore they are now stored in a map, instead of a list.
And the key for each of them is the pattern string used to create
them.

This change allows the hooks to be stored with more intuitive names
and encourage the users to access them.